Documentation and minor code cleanups
