+++ kanvas/README.md
@@ -57,6 +57,82 @@ dependencies {
 
 _Where `$platform` represents (should be replaced with) the desired platform dependency (e.g. `jvm`)._
 
+### Apple Setup
+
+Consuming Krayon for Apple targets (MacOS, iOS) requires additional setup, due to the inability to
+load multiple Kotlin libraries to Swift/Objective-C. The intended technique is for Krayon to be
+included directly by another Kotlin library which would contain shared application code, including
+shared Krayon drawing. That library would then export Krayon as part of its native library for use
+by iOS.
+
+#### Using CocoaPods Plugin
+
+Include the cocoapods plugin in your Kotlin module:
+
+```kotlin
+plugins {
+    kotlin("multiplatform")
+    kotlin("native.cocoapods")
+}
+```
+
+When defining your framework, make sure to export Krayon modules as needed:
+
+```kotlin
+cocoapods {
+    // ...
+    framework {
+        // ...
+        export("com.juul.krayon:$module:$currentVersion")
+    }
+}
+```
+
+#### Using Kotlin-Artifacts DSL
+
+A simpler, but experimental, approach involves using the [Kotlin Artifacts dsl](https://kotlinlang.org/docs/multiplatform-native-artifacts.html).
+
+```kotlin
+kotlinArtifacts {
+    Native.XCFramework("YourLibrary") {
+        targets(macosArm64, macosX64, iosArm64)
+        modes(DEBUG, RELEASE)
+        addModule("com.juul.krayon:$module:$currentVersion")
+    }
+}
+```
+
+#### Using Krayon
+
+The basic building block of Krayon for Apple is the `CGContextKanvas`, which allows drawing through
+Krayon's multiplatform interface with Core Graphics' native drawing primitives. Krayon assumes the
+`DeviceRGB` colorspace. For offscreen rendering, you can create your own:
+
+```swift
+// Define some constants. Here `scale` is used to handle retina 2x scaling
+val width = 400
+val height = 300
+let scale = 2
+
+// Create a context and pass it to Krayon
+var cgContext: CGContext = CGContext(
+    data: nil,
+    width: width * scale,
+    height: height * scale,
+    bitsPerComponent: 8,
+    bytesPerRow: 0,
+    space: CGColorSpaceCreateDeviceRGB(),
+    bitmapInfo: CGImageAlphaInfo.premultipliedLast.rawValue
+)!
+cgContext.scaleBy(x: scale, y: scale)
+// Note that Kanvas uses the unscaled size, which does not match the pixel count of the CGContext
+let kanvas = CGContextKanvas(ptr: &cgContext, width: width, height: height)
+```
+
+Drawing to your Krayon canvas directly from Swift isn't recommend -- the API is quite ugly. From here,
+you should call a Kotlin function that handles drawing for you. As an example, using the `element`
+module, you might simply have `yourRootElement.draw(kanvas: kanvas)`.
+
 [badge-android]: http://img.shields.io/badge/platform-android-6EDB8D.svg?style=flat
 [badge-ios]: http://img.shields.io/badge/platform-ios-CDCDCD.svg?style=flat
 [badge-js]: http://img.shields.io/badge/platform-js-F8DB5D.svg?style=flat
